feat: 统一回收站（双轨隔离与分组展示）
- 在 SQLite 数据库中独立开表 disk_trash 记录磁盘删除项，彻底不污染主元数据表 metadata
- 新建 DiskTrashService 异步代管磁盘模式下的删除、还原与永久抹除
- 实现双轨隔离与分组展示，在 UI 视图中绘制独立的“资源库 - 托管资产”与“目录导航 - 物理文件”分组标题框并进行物理排版隔离展示
- 打通 ContentPanel 与 CategoryPanel 一键清空和还原的双轨分流，彻底杜绝主线程 I/O 阻塞假死

+++ src/ui/ContentPanel.cpp
@@ -2011,7 +2011,7 @@ void ContentPanel::onCustomContextMenuRequested(const QPoint& pos) {
 
                 // 异步多线程执行物理安全删除与数据库记录清除 (双轨隔离)
                 (void)QtConcurrent::run([targetPaths, diskTrashItems, action, weakThis, weakProgress]() {
-                    int total = targetPaths.size() + diskTrashItems.size();
+                    int total = static_cast<int>(targetPaths.size() + diskTrashItems.size());
                     int count = 0;
 
                     // A. 处理常规项目 (资源库托管或非回收站物理文件)
